Fix ClockIO circular dep
